pyproject.toml
[tool.poetry]
name = "colrev"
version = "0.10.4"
description = "CoLRev: An open-source environment for collaborative reviews"
authors = ["Gerit Wagner <gerit.wagner@uni-bamberg.de>", "Julian Prester <julian.prester@sydney.edu.au>"]
license = "MIT"
readme = "README.md"
homepage="https://colrev.readthedocs.io/en/latest/"
repository = "https://github.com/CoLRev-Environment/colrev"
documentation = "https://colrev.readthedocs.io/en/latest/"
keywords = ["research", "reproducible research", "open science", "literature", "literature review", "systematic review", "systematic literature review"]
classifiers = [
   "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
include = ["CONTRIBUTING.md", "CHANGELOG.md", "LICENSE", "README.md", "colrev/py.typed", "colrev/.colrev_endpoints.json"]

[tool.poetry.scripts]
colrev = "colrev.ui_cli.cli:main"
colrev-hooks-check = "colrev.hooks.check:main"
colrev-hooks-format = "colrev.hooks.format:main"
colrev-hooks-report = "colrev.hooks.report:main"
colrev-hooks-share = "colrev.hooks.share:main"
colrev-hooks-update = "colrev.hooks.update:main"

[tool.poetry.dependencies]
python = ">=3.8, <4"
click-repl = "^0.3.0"
click = "^8.1.6"
requests = "^2.28.1"
requests-cache = "^0.9.6"
tqdm = "^4.64.1"
transitions = "^0.8.11"
pre-commit = "^2.20.0"
docker = "6.0.0"
GitPython= "^3.1.29"
beautifulsoup4 = "^4.11.1"
click_completion = "^0.5.2"
"zope.interface" = "^5.5.0"
alphabet-detector = "^0.0.7"
crossrefapi = "^1.5.0"
dacite = "^1.6.0"
daff = "^1.3.46"
dedupe = "2.0.23"
dictdiffer = "^0.9.0"
lingua-language-detector = ">1.3"
lxml = "^4.9.1"
nameparser = "^1.1.2"
pandas = "^1.5.0"
pandasql = "^0.7.3"
"pdfminer.six" = "^20220524"
psutil = "^5.9.2"
pybtex = "^0.24.0"
PyPDF2 = "^1.28.6"
PyYAML = "^6.0.0"
thefuzz = "^0.19.0"
pycountry = "^22.3.5"
glom = "^22.1.0"
Jinja2 = "^3.1.2"
dataclasses-jsonschema = "^2.15.3"
openpyxl = "^3.1.2"
pymupdf = "^1.22.0"
imagehash = "^4.3.1"
defusedxml = "^0.7.1"
dash = "^2.11.1"
levenshtein = "^0.21.0" # faster implementation of levenshtein distance (for thefuzz)
pyalex = "^0.10"
colrev-asreview = "^0.1.0"
m2r = "^0.3.1"
inquirer = "^3.1.3"
Sphinx = {version = "^5.2.3", optional = true }
sphinx-autodoc-typehints = {version = "^1.19.4", optional = true }
sphinx-click = {version = "^4.3.0", optional = true }
sphinx-rtd-theme = {version = "^1.1.1", optional = true }
"sphinxcontrib.datatemplates" = {version = "^0.9.2", optional = true }
sphinx_collapse = {version = "^0.1.2", optional = true }
repoze-sphinx-autointerface = {version = "^1.0.0", optional = true }
pylint = {version = "3.0.1", optional = true }
pytest = {version = "^7.2.1", optional = true }
coverage = {extras = ["toml"], version = "^7.3.2", optional = true }
types-click = {version = "^7.1.8", optional = true }
pytest-mock = {version = "^3.10.0", optional = true }
requests-mock = {version = "^1.10.0", optional = true }
feedparser = "^6.0.10"

[tool.poetry.extras]
docs = ["Sphinx", "sphinx-autodoc-typehints", "sphinx-click", "sphinx-rtd-theme", "sphinxcontrib.datatemplates", "sphinx_collapse", "repoze-sphinx-autointerface"]
dev = ["pylint", "pytest", "coverage", "types-click", "pytest-mock", "requests-mock"]

[tool.mypy]
python_version = 3.8
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "requests.*",
    "yaml.*",
    "pkg_resources.*",
]
ignore_missing_imports = true

[tool.pylint.MAIN]
extension-pkg-whitelist = "lxml.etree"
load-plugins = ["colrev.linter.colrev_direct_status_assign", "colrev.linter.colrev_missed_constant_usage"]

[tool.pylint.'MESSAGES CONTROL']
disable = ["R0801"]

[tool.coverage.report]
omit = [
    "colrev/ops/built_in/*", # omit reference implementation
    "/usr/*", # omit everything in /usr
    "*crossrefapi/crossref/*",
    "colrev/env/grobid_service.py",
    "/tmp/*",
    "tests/*",
]

[tool.coverage.run]
omit = [
    "colrev/ops/built_in/*", # omit reference implementation
    "/usr/*", # omit everything in /usr
    "*crossrefapi/crossref/*",
    "colrev/env/grobid_service.py",
    "/tmp/*",
    "tests/*",
]

[tool.ruff]
line-length = 160

[tool.ruff.per-file-ignores]
"**/{tests,docs}/*" = ["E501"]

[build-system]
# requires = ["poetry>=1.2.0b1"]
requires = ["poetry-core>=1.0.0", "cython<3.0"]
build-backend = "poetry.core.masonry.api"
